Addressed PR comments. * Added configuration parameters for http server host and port. * Renamed ethereum provider to provider. * Updated which errors are thrown when fetching JSON response. * Removed unneeded exception handling.
