+++ packages/buidler-core/src/internal/buidler-evm/jsonrpc/handler.ts
@@ -12,7 +12,7 @@ import {
   isValidJsonResponse,
   JsonRpcRequest,
   JsonRpcResponse
-} from "../../core/providers/http";
+} from "../../util/jsonrpc";
 import {
   BuidlerEVMProviderError,
   InternalError,
