+++ packages/buidler-core/src/internal/core/providers/http.ts
@@ -1,37 +1,18 @@
 import { EventEmitter } from "events";
 
+import {
+  FailedJsonRpcResponse,
+  JsonRpcRequest,
+  JsonRpcResponse,
+  parseJsonResponse
+} from "../../util/jsonrpc";
 import { BuidlerError } from "../errors";
 import { ERRORS } from "../errors-list";
-
-export interface JsonRpcRequest {
-  jsonrpc: string;
-  method: string;
-  params: any[];
-  id: number;
-}
-
-interface SuccessfulJsonRpcResponse {
-  jsonrpc: string;
-  id: number;
-  result: any;
-}
-
-interface FailedJsonRpcResponse {
-  jsonrpc: string;
-  id?: number;
-  error: {
-    code: number;
-    message: string;
-    data?: any;
-  };
-}
 
 interface ProviderError extends Error {
   code?: number;
   data?: any;
 }
-
-export type JsonRpcResponse = SuccessfulJsonRpcResponse | FailedJsonRpcResponse;
 
 function isErrorResponse(response: any): response is FailedJsonRpcResponse {
   return typeof response.error !== "undefined";
@@ -115,70 +96,4 @@ export class HttpProvider extends EventEmitter {
       id: this._nextRequestId++
     };
   }
-}
-
-export function parseJsonResponse(text: string): JsonRpcResponse {
-  try {
-    const json = JSON.parse(text);
-
-    if (!isValidJsonResponse(json)) {
-      throw new Error();
-    }
-
-    return json;
-  } catch (error) {
-    throw new BuidlerError(ERRORS.NETWORK.INVALID_JSON_RESPONSE, {
-      response: text
-    });
-  }
-}
-
-export function isValidJsonRequest(payload: any): boolean {
-  if (payload.jsonrpc !== "2.0") {
-    return false;
-  }
-
-  if (typeof payload.id !== "number" && typeof payload.id !== "string") {
-    return false;
-  }
-
-  if (typeof payload.method !== "string") {
-    return false;
-  }
-
-  if (payload.params !== undefined && !Array.isArray(payload.params)) {
-    return false;
-  }
-
-  return true;
-}
-
-export function isValidJsonResponse(payload: any) {
-  if (payload.jsonrpc !== "2.0") {
-    return false;
-  }
-
-  if (
-    typeof payload.id !== "number" &&
-    typeof payload.id !== "string" &&
-    payload.id !== undefined
-  ) {
-    return false;
-  }
-
-  if (payload.result === undefined && payload.error === undefined) {
-    return false;
-  }
-
-  if (payload.error !== undefined) {
-    if (typeof payload.error.code !== "number") {
-      return false;
-    }
-
-    if (typeof payload.error.message !== "string") {
-      return false;
-    }
-  }
-
-  return true;
 }
